LPS-55518 match MentionsTopHeadDynamicInclude
